feat(loadbalancing): Extract load_balancing builder tests into a separate file
