+++ gateway/internal/features/builder_test.go
@@ -6,7 +6,6 @@ package features_test
 
 import (
 	"context"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/telekom/controlplane/common/pkg/types"
@@ -18,138 +17,10 @@ import (
 	"github.com/telekom/controlplane/gateway/pkg/kong/client/mock"
 	"github.com/telekom/controlplane/gateway/pkg/kong/client/plugin"
 	"go.uber.org/mock/gomock"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func NewMockRoute() *gatewayv1.Route {
-	return &gatewayv1.Route{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.RouteSpec{
-			Realm: types.ObjectRef{
-				Name:      "realm",
-				Namespace: "default",
-			},
-			PassThrough: false,
-			Upstreams: []gatewayv1.Upstream{
-				{
-					Url: "http://upstream.url:8080/api/v1",
-					// Default is used for Weight
-					Scheme: "http",
-					Host:   "upstream.url",
-					Port:   8080,
-					Path:   "/api/v1",
-				},
-			},
-			Downstreams: []gatewayv1.Downstream{
-				{
-					Host:      "downstream.url",
-					Port:      8080,
-					Path:      "/test/v1",
-					IssuerUrl: "issuer.url",
-				},
-			},
-		},
-	}
-}
-
-func NewMockConsumeRoute(routeRef types.ObjectRef) *gatewayv1.ConsumeRoute {
-	return &gatewayv1.ConsumeRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-consumer",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.ConsumeRouteSpec{
-			ConsumerName: "test-consumer-name",
-			Route:        routeRef,
-		},
-	}
-}
-
-func NewMockRealm() *gatewayv1.Realm {
-	return &gatewayv1.Realm{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-realm",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.RealmSpec{
-			Url:       "https://realm.url",
-			IssuerUrl: "https://issuer.url",
-			DefaultConsumers: []string{
-				"gateway",
-				"test",
-			},
-		},
-	}
-}
-
-func NewMockGateway() *gatewayv1.Gateway {
-	return &gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-gateway",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			Admin: gatewayv1.AdminConfig{
-				ClientId:     "admin",
-				ClientSecret: "topsecret",
-				IssuerUrl:    "https://issuer.url",
-				Url:          "https://admin.test.url",
-			},
-		},
-	}
-}
-
-func NewLoadBalancingUpstreams(isProxyRoute bool) []gatewayv1.Upstream {
-	var issuerUrl string
-	if isProxyRoute {
-		issuerUrl = "https://upstream.issuer.url"
-	} else {
-		issuerUrl = ""
-	}
-
-	return []gatewayv1.Upstream{
-		{
-			Url:       "http://upstream.url:8080/api/v1",
-			Weight:    2,
-			Scheme:    "http",
-			Host:      "upstream.url",
-			Port:      8080,
-			Path:      "/api/v1",
-			IssuerUrl: issuerUrl,
-		},
-		{
-			Url:       "http://upstream2.url:8080/api/v1",
-			Weight:    1,
-			Scheme:    "http",
-			Host:      "upstream2.url",
-			Port:      8080,
-			Path:      "/api/v1",
-			IssuerUrl: issuerUrl,
-		},
-	}
-}
-
 var _ = Describe("FeatureBuilder", Ordered, func() {
-	var mockCtrl *gomock.Controller
-	BeforeAll(func() {
-		mockCtrl = gomock.NewController(GinkgoT())
-	})
-
 	Context("Registering", Ordered, func() {
-
-		var route *gatewayv1.Route
-		var realm *gatewayv1.Realm
-		var gateway *gatewayv1.Gateway
-
-		BeforeAll(func() {
-			route = NewMockRoute()
-			realm = NewMockRealm()
-			gateway = NewMockGateway()
-		})
-
 		It("should be registered", func() {
 			kc := mock.NewMockKongClient(mockCtrl)
 
@@ -167,22 +38,8 @@ var _ = Describe("FeatureBuilder", Ordered, func() {
 	})
 
 	Context("Applying and Creating", Ordered, func() {
-
 		var ctx = context.Background()
-		var mockKc *mock.MockKongClient
-
-		var route *gatewayv1.Route
-		var realm *gatewayv1.Realm
-		var gateway *gatewayv1.Gateway
-
-		BeforeAll(func() {
-			route = NewMockRoute()
-			realm = NewMockRealm()
-			gateway = NewMockGateway()
-
-			ctx = contextutil.WithEnv(ctx, "test")
-		})
-
+		ctx = contextutil.WithEnv(ctx, "test")
 		BeforeEach(func() {
 			mockKc = mock.NewMockKongClient(mockCtrl)
 		})
@@ -357,88 +214,6 @@ var _ = Describe("FeatureBuilder", Ordered, func() {
 		It("should correctly apply the RateLimit feature", func() {
 			// TBD
 			Expect(true).To(BeTrue())
-		})
-
-		It("should apply the LoadBalancing feature", func() {
-			loadBalancingRoute := route.DeepCopy()
-			loadBalancingRoute.Spec.Upstreams = NewLoadBalancingUpstreams(false)
-			builder := features.NewFeatureBuilder(mockKc, loadBalancingRoute, realm, gateway)
-			builder.EnableFeature(feature.InstanceLoadBalancingFeature)
-
-			mockKc.EXPECT().CreateOrReplaceRoute(ctx, loadBalancingRoute, gomock.Any()).Return(nil).Times(1)
-			mockKc.EXPECT().CreateOrReplacePlugin(ctx, gomock.Any()).Return(nil, nil).Times(1)
-			mockKc.EXPECT().CleanupPlugins(ctx, gomock.Any(), gomock.Any()).Return(nil).Times(1)
-
-			By("building the features")
-			err := builder.Build(ctx)
-			Expect(err).ToNot(HaveOccurred())
-
-			b, ok := builder.(*features.Builder)
-			Expect(ok).To(BeTrue())
-
-			By("checking the jumper config")
-			jumperConfig := b.JumperConfig()
-
-			By("Checking that JumperConfig contains both upstreams with weights")
-			Expect(jumperConfig).NotTo(BeNil())
-			Expect(jumperConfig.LoadBalancing).NotTo(BeNil())
-			Expect(jumperConfig.LoadBalancing.Servers).To(HaveLen(2))
-			Expect(jumperConfig.LoadBalancing.Servers[0].Upstream).To(Equal("http://upstream.url:8080/api/v1"))
-			Expect(jumperConfig.LoadBalancing.Servers[0].Weight).To(Equal(2))
-			Expect(jumperConfig.LoadBalancing.Servers[1].Upstream).To(Equal("http://upstream2.url:8080/api/v1"))
-			Expect(jumperConfig.LoadBalancing.Servers[1].Weight).To(Equal(1))
-
-			By("checking the request-transformer plugin")
-			rtPlugin, ok := b.Plugins["request-transformer"].(*plugin.RequestTransformerPlugin)
-			Expect(ok).To(BeTrue())
-
-			By("checking the request-transformer plugin config")
-			Expect(rtPlugin.Config.Append.Headers).To(BeNil())
-			Expect(rtPlugin.Config.Remove.Headers).To(BeNil())
-		})
-
-		It("should apply the LastMileSecurity and the LoadBalancing feature", func() {
-			loadBalancingRoute := route.DeepCopy()
-			loadBalancingRoute.Spec.PassThrough = false
-			loadBalancingRoute.Spec.Upstreams = NewLoadBalancingUpstreams(false)
-
-			builder := features.NewFeatureBuilder(mockKc, loadBalancingRoute, realm, gateway)
-			builder.EnableFeature(feature.InstanceLastMileSecurityFeature)
-			builder.EnableFeature(feature.InstanceLoadBalancingFeature)
-
-			mockKc.EXPECT().CreateOrReplaceRoute(ctx, loadBalancingRoute, gomock.Any()).Return(nil).Times(1)
-			mockKc.EXPECT().CreateOrReplacePlugin(ctx, gomock.Any()).Return(nil, nil).Times(1)
-			mockKc.EXPECT().CleanupPlugins(ctx, gomock.Any(), gomock.Any()).Return(nil).Times(1)
-
-			By("building the features")
-			err := builder.Build(ctx)
-			Expect(err).ToNot(HaveOccurred())
-
-			b, ok := builder.(*features.Builder)
-			Expect(ok).To(BeTrue())
-
-			By("checking the jumper config")
-			jumperConfig := b.JumperConfig()
-
-			By("Checking that JumperConfig contains both upstreams with weights")
-			Expect(jumperConfig).NotTo(BeNil())
-			Expect(jumperConfig.LoadBalancing).NotTo(BeNil())
-			Expect(jumperConfig.LoadBalancing.Servers).To(HaveLen(2))
-			Expect(jumperConfig.LoadBalancing.Servers[0].Upstream).To(Equal("http://upstream.url:8080/api/v1"))
-			Expect(jumperConfig.LoadBalancing.Servers[0].Weight).To(Equal(2))
-			Expect(jumperConfig.LoadBalancing.Servers[1].Upstream).To(Equal("http://upstream2.url:8080/api/v1"))
-			Expect(jumperConfig.LoadBalancing.Servers[1].Weight).To(Equal(1))
-
-			By("checking the request-transformer plugin")
-			rtPlugin, ok := b.Plugins["request-transformer"].(*plugin.RequestTransformerPlugin)
-			Expect(ok).To(BeTrue())
-
-			By("checking the request-transformer plugin config")
-			Expect(rtPlugin.Config.Append.Headers.Contains("remote_api_url")).To(BeFalse())
-			Expect(rtPlugin.Config.Append.Headers.Contains("api_base_path")).To(BeTrue())
-			Expect(rtPlugin.Config.Append.Headers.Contains("access_token_forwarding")).To(BeTrue())
-			Expect(rtPlugin.Config.Append.Headers.Contains(plugin.JumperConfigKey)).To(BeTrue())
-			Expect(rtPlugin.Config.Remove.Headers.Contains("consumer-token")).To(BeTrue())
 		})
 
 		// TBD other features
